pyproject.toml
[project]
name = "mba-ct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.49",
    "cryptography>=46.0.2",
    "fastapi>=0.118.3",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pymysql>=1.1.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.0",
    "strands-agents>=1.12.0",
    "strands-agents-tools>=0.2.10",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
]


[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
web = ["jinja2"]


[project.scripts]
# CLI tools you already had
mba-ingest = "MBA.cli:main"
mba-producer = "MBA.microservices.producer:main"
mba-worker = "MBA.microservices.worker:main"
mba-api = "MBA.microservices.api:run_server"
mba-db = "MBA.database.cli:main"

# NEW: Streamlit launcher (calls MBA.streamlit_app:main)
mba-app = "MBA.app_launcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/MBA"]
